+++ articles/serializers.py
@@ -7,7 +7,7 @@ from articles.models import (
     ArticleImages, 
 )
 from users.models import Evaluations
-from users.serializers import EvaluationSerializer
+from profiles.serializers import EvaluationSerializer
 from django.contrib.auth import get_user_model
 
 User = get_user_model()
