+++ users/urls.py
@@ -10,13 +10,7 @@ urlpatterns = [
     path('token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),  # Refresh 토큰 갱신
     path('delete/', views.CustomDeleteUserView.as_view(), name='delete_user'), 
     path('profile/', views.UserProfileView.as_view(), name='user_profile'),
-    path('users/matching/', views.MatchingPageView.as_view(), name='user_matching'),
-    path('users/recommendations/', views.UserRecommendationView.as_view(), name='user_recommendations'), # 유저 추천 기능
-    path('users/rankings/', views.MannerRankingView().as_view(), name='user-rankings'), # 매너 랭킹
-    path('search/', views.SearchPageView.as_view(), name='user-search'), # 유저 검색
-    path('users/<str:username>/', views.UserDetailView().as_view(), name='user_detail'), # 특정 유저 검색 기능
     path('discordlogin/', views.discordLoginView.as_view(), name='discord-login'), #디스코드로 로그인
-    path('', views.indexView.as_view(), name='user_index'),
     
     # Riot 로그인 관련 URL
     # path('riot/', views.login_with_riot, name='login_with_riot'),  # auth/riot/
