+++ wpgg/urls.py
@@ -23,4 +23,5 @@ urlpatterns = [
     path('auth/', include('users.urls')),  # 사용자 정의 URL 설정
     path('articles/', include('articles.urls')),  # articles URL 설정
     path('api/chat/', include('chats.urls')),  # chat URL 설정
+    path('api/profile/', include('profiles.urls')),
 ]
